Bot replies with the standup text
