--- .env.example
@@ -1,2 +1,0 @@
-DISCORD_TOKEN=""
-APP_ID=""
