--- config/code.json
@@ -1,3 +1,0 @@
-{
-    "redeemableCodes": []
-}
