--- config/database.yml
@@ -1,10 +1,0 @@
-####################################################################
-#                        Database Settings                         #
-####################################################################
-
-user: user
-dbname: bot
-password: pass
-host: localhost
-port: 40041
-
