--- config/role.json
@@ -1,3 +1,0 @@
-{
-    "allowedRoles": []
-}
