--- data/log.txt
@@ -1,2 +1,0 @@
-Name     |  message
-FriedRice hello my name is
