--- events/cb_ready.js
@@ -1,11 +1,0 @@
-const { CarrotboardStorage } = require("../lib/carrotboard");
-
-module.exports = {
-    name: "ready",
-    once: true,
-    execute(client) {
-        const cbStorage = new CarrotboardStorage(client);
-        global.cbStorage = cbStorage;
-    },
-};
-
